기록이 존재하는 서브카테고리 제거 시 에러 처리

+++ springboot-back/src/main/java/com/health/springbootback/controller/DataController.java
@@ -227,6 +227,11 @@ public class DataController {
 
             if(!categoryService.existSubCategoryByEid(eid))
                 return ResponseEntity.ok().body(new MsgResponseDto(false, "SubCategory가 존재하지 않습니다"));
+            System.out.println("1");
+            if(recordService.exeistRecordBySubCategory(eid))
+                return ResponseEntity.ok().body(new MsgResponseDto(false, "SubCategory가 내에 Record 존재합니다"));
+            System.out.println("2");
+
             categoryService.deleteSubCategory(eid);
             return ResponseEntity.ok().body(new MsgResponseDto(true, "SubCategory 삭제"));
         } catch(HttpStatusCodeException | JsonProcessingException e) {
